refactor: move triggers to models
